% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diversity.R
\name{alpha_diversity}
\alias{alpha_diversity}
\title{Title Wrapper function to analyse alpha diversity.}
\usage{
alpha_diversity(
  physeq,
  measure = "Shannon",
  x,
  group = NULL,
  plot_type = "boxplot",
  color_vector = c("brown", "darkgreen", "orange", "violet"),
  stat = FALSE,
  check_depth = F,
  size = 1.5
)
}
\arguments{
\item{physeq}{A phyloseq object}

\item{measure}{The type of alpha diversity measure you wish to use. These are directly taken from \link{estimate_richness}.}

\item{x}{String character to display on X axis}

\item{group}{String character to color/fill by.}

\item{plot_type}{For the moment the function accepts "line" or "boxplot". Default = "boxplot"}

\item{stat}{Logical. Show stats or not. Default= FALSE.}

\item{check_depth}{Logical. Only for integers count, doesn't work with relative abundance. Will create two plots :
\itemize{
\item a first one being a boxplot based ggplot with depth as size of the jitters
\item a second one being a dotplot with alpha diversity measure on Y and the depth on X.
}}

\item{size}{Numeric value for the dot size and line width.}
}
\value{
Return a ggplot graphic.
}
\description{
This function will rapidly plot alpha diversity in ggplot2.
}
\examples{
data(enterotype)

alpha_diversity(enterotype, measure="Shannon", x="Enterotype", group="SeqTech", plot_type="boxplot")
# same analysis but with check_depth
alpha_diversity(GlobalPatterns, measure="Shannon", x="SampleType", plot_type="boxplot", check_depth=T)
}
