% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity.R
\name{beta_diversity}
\alias{beta_diversity}
\title{Plot alpha diversity with numerous aspect.}
\usage{
beta_diversity(
  physeq,
  dist = "wunifrac",
  nf = 5,
  method = "PCoA",
  group = NULL,
  color_vector = c("cyan4", "brown", "deepskyblue", "black", "red"),
  factor_to_plot = NULL,
  second_factor_to_plot = NULL,
  lwd = 1,
  cpoint = 2,
  clabel = 1,
  addaxes = F,
  axesell = T,
  csub = 2,
  font = 2,
  pch = 20,
  species = F,
  permanova = F,
  where = "topleft",
  ...
)
}
\arguments{
\item{physeq}{(Required). A phyloseq object to analyse your data.}

\item{method}{Computation method to display beta-diversity. Default : \code{"PCoA"}}

\item{group}{A character string from your metadata. Select one variable you want to segregate your data by.}

\item{color_vector}{A color vector by which you want plot. Make sure the colovector is sufficient for your parameters.
Default : c("cyan4","brown","deepskyblue", "black","red")}

\item{factor_to_plot}{(Optionnal). A character string you want as title.}

\item{second_factor_to_plot}{(Optionnal). A second factor.}

\item{lwd}{Line width.}

\item{cpoint}{Point size}

\item{clabel}{Label size.}

\item{addaxes}{Plot axis or not. Logical.}

\item{axesell}{Plot axesell. See \link[ade4:s.class]{ade4::s.class}.}

\item{csub}{Title size.}

\item{distance}{A character string. The default is "wunifrac" from the phyloseq package.}
}
\description{
Plot alpha diversity with numerous aspect.
}
