% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_distribution.R
\name{check_distribution}
\alias{check_distribution}
\title{Title}
\usage{
check_distribution(df, values, factor, normalisation = "none")
}
\arguments{
\item{df}{Dataframe to be used}

\item{values}{Values (numeric) to be used}

\item{factor}{Factor use for plotting}

\item{normalisation}{Normalisation method. So far the function accepts "log", "clr", "scaling", "sum scaling", "sqrt", "pareto", "crt", "vast", "level", "median centering", "gm centering" and "all".
"all" will give a facetted plot with all method depicted.}
}
\value{
Returns a plot with the given normalisation method
}
\description{
Title
}
\examples{
TBD
}
