% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_length.R
\name{check_length}
\alias{check_length}
\title{Check the length of reads in fastq.}
\usage{
check_length(fastq, names, cores = 1)
}
\arguments{
\item{fastq}{List of fastq}

\item{names}{Names of the samples}

\item{cores}{Number of threads to use. Default=1}
}
\value{
TBD

A dataframe with the length of sequences for each fastq.
}
\description{
Check the length of reads in fastq.
}
\examples{
No example
}
