% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_variance_plot.R
\name{circular_variance_plot}
\alias{circular_variance_plot}
\title{Make circular plots for variance analysis.}
\usage{
circular_variance_plot(
  variance,
  adjust = 1,
  col = c("brown", "darkgreen", "orange", "violet"),
  show.percentage = F,
  label.size = 5,
  text.size = 5
)
}
\arguments{
\item{variance}{Object returned by the function \link{calculate_variance}.}

\item{adjust}{Value to adjust the labels}

\item{show.percentage}{Logical. Show percentage for each features.}

\item{label.size}{Size of the labels.}

\item{text.size}{Size of the labels}
}
\value{
A ggplot2 graphic.
}
\description{
A ggplot2 based function to plot variance explained by factors in a circular manner.
}
\examples{

var = calculate_variance(metabolomic[,1:4], features = metabolomic[,5:217], cores = 5)
circular_variance_plot(variance, adjust=0.4, show.percentage = T)
}
