% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sra_metadata.R
\name{clean_sra_metadata}
\alias{clean_sra_metadata}
\title{Title}
\usage{
clean_sra_metadata(dir, regex_match, output_path = "./curated_metadata")
}
\arguments{
\item{output_path}{}
}
\description{
Title
}
