% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_IgAseq.R
\name{collapse_IgAseq}
\alias{collapse_IgAseq}
\title{Collapse IgASeq}
\usage{
collapse_IgAseq(IgAseq)
}
\arguments{
\item{IgAseq}{Object returned by the slide_z() function}
}
\value{
IgAseq object. Refers to \link{slide_z} for more information on the IgAseq object.
}
\description{
Collapse the object returned by slide_z(). The slide_z() function needs to be performed in a for loop or lapply and be collapsed into a single object.
}
\examples{
URL to come.
}
