% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_beta_dispersion.R
\name{constrained_beta_dispersion}
\alias{constrained_beta_dispersion}
\title{Make a constrained analysis of beta diversity.}
\usage{
constrained_beta_dispersion(
  physeq,
  axis_x = 1,
  axis_y = 2,
  model,
  dist = "bray",
  nf = 5,
  method = "CCA",
  group = NULL,
  color_vector = c("cyan4", "brown", "deepskyblue", "black", "red"),
  legend_title = NULL,
  lwd = 1,
  conf = 0.9,
  cex = 2,
  font = 2,
  pch = 20,
  draw = "lines",
  ylimits = "auto",
  xlimits = "auto",
  text = F,
  ncol = 1,
  species = F,
  x.intersp = 1,
  y.intersp = 0.5,
  permanova = F,
  where = "topleft",
  inset = 0.2,
  boxplot = T,
  ...
)
}
\arguments{
\item{physeq}{A phyloseq object.}

\item{axis_x}{Which component for axis x. Default = 1}

\item{axis_y}{Which component for axis y. Default = 2}

\item{model}{Character vector to create a model. This is related to the functions rda, cca and capscale.}

\item{dist}{Either a distance mecthod given as string or a distance object.
Default= "bray".}

\item{nf}{Number of component to keep. Default= 5.}

\item{method}{The mathematical method to reduce the dimensions. Currently this function supports CCA, RDA and dbRDA}

\item{group}{String defining the groups you want to plot on the beta diversity.}

\item{color_vector}{Optionnal. A color vector you wish to use for plotting}

\item{legend_title}{Optionnal. A legend title.}

\item{lwd}{Line width}

\item{conf}{Interval of confidence for the ellipses. Default= 0.9.}

\item{cex}{Text size}

\item{font}{Font type}

\item{pch}{Shape type}

\item{draw}{Either draw lines or polygon. Default = "lines".}

\item{ylimits}{Limits for y axis}

\item{xlimits}{Limits for x axis}

\item{text}{Boolean. Plot the names of the group on the graph.}

\item{ncol}{Number of columns for the legend. Default = 1.}

\item{species}{Not implemented.}

\item{x.intersp}{character interspacing factor for horizontal (x) spacing between symbol and legend text.}

\item{y.intersp}{vertical (y) distances (in lines of text shared above/below each legend entry). A vector with one element for each row of the legend can be used.}

\item{permanova}{Make the permanova analysis. Default = FALSE.}

\item{where}{Where you want the permanova result.}

\item{inset}{Permanova legend adjustment see \link{legend}}

\item{boxplot}{Draw boxplot for each axis. Default = T}
}
\description{
A wrapper function to use cca, rda and capscale in the same manner as beta_dispersion.
}
\examples{
res= constrained_beta_dispersion(enterotype,
                                 model= "SeqTech+Gender+Nationality+Age+ClinicalStatus",
                                 group="Enterotype", method = "CCA",
                                 boxplot =T,
                                 text=T,
                                 color_vector = tol21rainbow)
#res
}
