% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate_data.R
\name{cross_validate_data}
\alias{cross_validate_data}
\title{Cross validate the training data.}
\usage{
cross_validate_data(
  obj,
  y,
  model = "plsda",
  repetition = 1000,
  cores = 1,
  ncomp = 10,
  number = 10,
  repeats = 3,
  prop = 0.5,
  ...
)
}
\arguments{
\item{obj}{An object of returned by \code{\link[=screen_models]{screen_models()}}}

\item{y}{Character vector of the desire variable to test.}

\item{model}{Type of model to use}

\item{repetition}{The number of repetition}

\item{cores}{The number of cores to use}

\item{ncomp}{The number of components you want to keep in plsda or splsda}

\item{number}{Either the number of folds or number of resampling iterations}

\item{repeats}{For repeated k-fold cross-validation only: the number of complete sets of folds to compute. Default = 3}

\item{prop}{The proportion of the variables to use in the training model. Default = 0.5.}

\item{...}{}
}
\value{
A list containing of ml class object.
\item{res_model}{The \code{\link[caret:train]{caret::train()}} value}
\item{res_prediction}{The prediction of the model}
\item{res_conf_matrix}{The confusion matrix and associated values, see \code{\link[caret:confusionMatrix]{caret::confusionMatrix()}}.
\item{data}{The data used in each models}
}
\description{
Cross validate the training data.
}
\examples{
TBD.
}
