% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse_me.R
\name{ellipse_me}
\alias{ellipse_me}
\title{Title}
\usage{
ellipse_me(df, confidence_interval)
}
\arguments{
\item{df}{dataframe returned by \link{seq_table}}

\item{confidence_interval}{interval of confidence}
}
\value{
An ellipse dataframe
}
\description{
Title
}
\examples{
No example
}
