% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fastq.R
\name{filter_fastq}
\alias{filter_fastq}
\title{Title}
\usage{
filter_fastq(
  fastq_list,
  filt_list,
  cutting_param,
  maxEE,
  trimleft = 0,
  cores,
  ...
)
}
\arguments{
\item{fastq_list}{List of fastq returned by \link{list_fastq}.}

\item{filt_list}{List of filtered fastq names givent by \link{filt_list}.}

\item{cutting_param}{Equivalent to \code{truncLen} in \link{filterAndTrim}.}

\item{maxEE}{Equivalent to \code{maxEE} in \link{filterAndTrim}.}

\item{trimleft}{Left length}

\item{cores}{Number of cores to used (Windows user cannot change this argument)}
}
\description{
Title
}
