% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_seq.R
\docType{class}
\name{IgASeq-class}
\alias{IgASeq-class}
\alias{ig_seq}
\title{Create a S4 object for the IgASeq data}
\value{
A class IgASeq
}
\description{
Create a S4 object for the IgASeq data
}
\section{Slots}{

\describe{
\item{\code{ig_seq_all}}{data.frame.}

\item{\code{ig_up}}{data.frame.}

\item{\code{ig_down}}{data.frame.}

\item{\code{ellipse_data}}{data.frame.}

\item{\code{taxa}}{data.frame.}
}}

\examples{
.
}
