% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_ratio.R
\name{log_ratio}
\alias{log_ratio}
\title{Preprocess relative abundance.}
\usage{
log_ratio(
  df,
  positive_sorted_sample,
  negative_sorted_sample,
  second_negative_sample,
  zero_treatment
)
}
\arguments{
\item{df}{A dataframe given by \code{\link[=seq_table]{seq_table()}}.}

\item{positive_sorted_sample}{Positive sorted sample pattern.}

\item{negative_sorted_sample}{First negative sorted sample pattern.}

\item{second_negative_sample}{Second negative sorted sample pattern.}

\item{zero_treatment}{Zero treatment to apply, see \code{\link[=random_generation]{random_generation()}}, \code{\link[=no_zero]{no_zero()}} or \code{\link[=minimum_count]{minimum_count()}}.}
}
\value{
Return a dataframe for each sample containing :
\item{log10_abundance}{The \eqn{log10(pos*neg1)} ratio}
\item{log2_ratio}{The \eqn{log2(pos1/neg1)}}
\item{log2_neg_ratio} {The \eqn{log2(neg1/neg2)}}
\item{log10_neg_abundance}{The \eqn{log10(neg1*neg2)}}
\item{taxa} {A collapsed taxonomy}
}
\description{
Process zero values and produce ratios for \code{\link[=slide_z]{slide_z()}}. This function requires a \code{zero_treatment} value,
make sure to use \link{neg_dispersion} to see which method is the most convenient and accurate for your data.
}
\examples{
No example
}
