% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_and_remove_chimera.R
\name{merge_and_remove_chim}
\alias{merge_and_remove_chim}
\title{Merge reads and remove chimera}
\usage{
merge_and_remove_chim(dd, filt_fastq_list, cores, ...)
}
\arguments{
\item{...}{}
}
\description{
Merge reads and remove chimera
}
