% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\docType{class}
\name{ml-class}
\alias{ml-class}
\alias{ml}
\title{Create a S4 object for the machine learning results.}
\value{
A class ml
}
\description{
Create a S4 object for the machine learning results.
}
\section{Slots}{

\describe{
\item{\code{model}}{List of models returned}

\item{\code{pred}}{List of predictions}

\item{\code{conf}}{List of confusion matrix and the attached statistics}

\item{\code{data}}{List of the data used to generate this object}
}}

\examples{
.
}
