% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg_dispersion.R
\name{neg_dispersion}
\alias{neg_dispersion}
\title{Check the dispersion of your samples and the impact of the zero_treatment.}
\usage{
neg_dispersion(
  df,
  positive_sorted_sample = "pos",
  negative_sorted_sample = "neg1",
  second_negative_sample = "neg2",
  type = "facet_all_three"
)
}
\arguments{
\item{df}{A element of the \code{\link[=seq_table]{seq_table()}} returned object. You can run this function using a loop or lapply to make the plots for every samples.}

\item{positive_sorted_sample}{The pattern of the positive fraction. Default = "pos"}

\item{negative_sorted_sample}{The pattern of the first negative fraction. Default = "neg1"}

\item{second_negative_sample}{The pattern of the second negative fraction. Default = "neg2"}

\item{type}{Either a facet of the three possibilities as a single panel: meaning the \code{\link[=random_generation]{random_generation()}},
the \code{\link[=minimum]{minimum()}} or \code{\link[=no_zero]{no_zero()}} functions as a single panel
or superposed : the \eqn{pos/neg1} ratio for each way of dealing with zero values on one panel and all the \eqn{neg1/neg2} ratio on an other panel.}
}
\value{
Return a ggplot2 graph.
\item{facet_all_three}{For the facet_all_three : the black dots will represent the \code{\link[=no_zero]{no_zero()}}, the green dots the \code{\link[=random_generation]{random_generation()}}
and the red dots the \code{\link[=minimum_count]{minimum_count()}}.}
\item{superposed}{For the superposed : the grey dots represent the \eqn{neg1/neg2} and the red dots represent \eqn{pos/neg1}.}
}
\description{
Check the dispersion of your samples and the impact of the zero management you want to perform.
There are three possibilities \code{\link[=random_generation]{random_generation()}}, \code{\link[=no_zero]{no_zero()}} and \code{\link[=minimum_count]{minimum_count()}}.
It will produce a list of plots were you will find the \eqn{neg1/neg2} ratio and the \eqn{pos/neg1} ratio.
This is particularly important to check once you're working on a new set of data.
Take into account that
\item random_generation will be usefull for samples with a few taxa. If you samples are from healthy adults you might want to use
}
\examples{
Check url to come.
}
