% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_analysis.R
\name{permutation_analysis}
\alias{permutation_analysis}
\title{Title}
\usage{
permutation_analysis(
  data,
  formula,
  clinical_data,
  distance = "euclidean",
  by = "margin"
)
}
\arguments{
\item{data}{A dataframe}

\item{formula}{The right side of the formula you want to apply to permanova}

\item{clinical_data}{Column names or number that are metadata}

\item{distance}{Distance to use, using vegan distances}

\item{by}{Either margin or terms}
}
\value{
Returns a plot and the p-values of the permanova.
}
\description{
Title
}
