% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_heatmap.R
\name{phylo_heatmap}
\alias{phylo_heatmap}
\title{A function to plot heatmap nicely, using a phylogenetic distance.}
\usage{
phylo_heatmap(
  physeq = physeq,
  labels = NULL,
  top = 30,
  factor_to_plot = NULL,
  distance = "wunifrac",
  taxa_rank = "Genus",
  split = NULL,
  color_vector = c("white", "#88CCAA", "#771122"),
  taxa_names_par = gpar(fontsize = 15, fontface = "bold.italic"),
  label_names_par = gpar(fontsize = 15, fontface = "bold")
)
}
\arguments{
\item{physeq}{(Required). A phyloseq object.}

\item{top}{The top number of ASV you want to represent. Default is  \code{30}. It's usually the best number to plot.}

\item{factor_to_plot}{(Optional) Factor for the title.}

\item{distance}{Phylogenetic dinstance to use as dendrogramm. See \code{\link{beta_diversity}}.}

\item{taxa_rank}{The taxa rank at which you want to agregate abundances. Default is \code{"Genus"}.}

\item{split}{(Careful !). Numeric number by which you want the dendrogramm to be split.
Work only with ComplexHeatmap version 2.2.0.}

\item{color_vector}{A character string for color plottinh. Default is \code{c("white","#88CCAA","#771122")}.}

\item{taxa_names_par}{Graphical parameters. Default is \code{gpar(fontsize= 15, fontface= "bold.italic")}}

\item{label_names_par}{Graphical paramaters. Default is \code{gpar(fontsize=15, fontface="bold")}}
}
\description{
A function to plot heatmap nicely, using a phylogenetic distance.
}
\examples{
library(phyloseq)
data(enterotype)
phylo_heatmap(physeq=enterotype, top=30, labels=NULL, distance="bray",
taxa_rank="Family")
}
