% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_constrained_reduction.R
\name{plot_constrained_reduction}
\alias{plot_constrained_reduction}
\title{Make a constrained reduction of your dataset.}
\usage{
plot_constrained_reduction(
  mat,
  cat_data,
  axis_x = 1,
  axis_y = 2,
  model,
  nf = 5,
  method = "CCA",
  group = NULL,
  permanova = F,
  color_vector = c("cyan4", "brown", "deepskyblue", "black", "red"),
  legend_title = NULL,
  lwd = 1,
  conf = 0.9,
  cex = 2,
  font = 2,
  pch = 20,
  draw = "lines",
  ylimits = "auto",
  xlimits = "auto",
  text = F,
  ncol = 1,
  x.intersp = 1,
  y.intersp = 0.5,
  where = "topleft",
  inset = 0.2,
  boxplot = T,
  ...
)
}
\arguments{
\item{mat}{A numeric matrix with samples as rows}

\item{cat_data}{The categorical datasets containing the factors to build the model}

\item{axis_x}{Which component for axis x. Default = 1}

\item{axis_y}{Which component for axis y. Default = 2}

\item{model}{Specify here the model to be used for regression}

\item{nf}{Number of components}

\item{method}{The mathematical method to reduce the dimensions. Currently this function supports CCA, RDA and dbRDA}

\item{group}{String defining the groups you want to plot on the beta diversity.}

\item{permanova}{Make the permanova analysis. Default = FALSE.}

\item{color_vector}{Optionnal. A color vector you wish to use for plotting}

\item{legend_title}{Optionnal. A legend title.}

\item{lwd}{Line width}

\item{cex}{Text size}

\item{font}{Font type}

\item{pch}{Shape type}

\item{ylimits}{Limits for Y axis, if not specified they will be automatically set}

\item{xlimits}{Limits for X axis, if not specified they will be automatically set}

\item{text}{Plot labels or not.}

\item{ncol}{Number of columns for the legend. Default = 1.}

\item{x.intersp}{character interspacing factor for horizontal (x) spacing between symbol and legend text.}

\item{y.intersp}{vertical (y) distances (in lines of text shared above/below each legend entry). A vector with one element for each row of the legend can be used.}

\item{where}{Where you want the permanova result.}

\item{inset}{Permanova legend adjustment see \link{legend}}

\item{boxplot}{Draw boxplot for each axis. Default = T}

\item{...}{}
}
\description{
Currently the function supports only the constrained reduction around categorical data. It's not possible yet to give two different Omics datasets to find correlation between them.
}
