% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corr_heatmap.R
\name{plot_corr_heatmap}
\alias{plot_corr_heatmap}
\title{Plot a correlation heatmap}
\usage{
plot_corr_heatmap(
  X,
  Y,
  method = "spearman",
  cutoff = 0,
  cluster = F,
  mid = "white",
  low = "lightcyan4",
  high = "lightsalmon4",
  ratio = 0.3,
  return_df = F
)
}
\arguments{
\item{X}{Numeric matrix to be correlated to the Y matrix}

\item{Y}{Numeric matrix to be correlated to the X matrix}

\item{method}{Statistical method used to}

\item{cutoff}{Minimum correlation value to plot}

\item{return_df}{}
}
\description{
A heatmap of correlation coefficient based on matrices.
}
