% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot the distribution of your data.}
\usage{
plot_distribution(
  mat,
  top = 50,
  selection_var = "mean.abun",
  col = c("orange", "brown"),
  return_df = F
)
}
\arguments{
\item{mat}{A matrix or data.frame with samples as rows}

\item{top}{The top N features to keep. Default = 50}

\item{selection_var}{Which variable is to be used for selecting the top features. Default = "mean.abundance"}

\item{col}{Color vector}

\item{return_df}{Logical. Return the data.frame used to plot. Default = T.}
}
\description{
A barplot showing the mean abundance and the variance for each features.
}
