% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loadings.R
\name{plot_loadings}
\alias{plot_loadings}
\title{Title}
\usage{
plot_loadings(ml_obj, top = 20)
}
\arguments{
\item{ml_obj}{An object of class ml or a list of ml class}
}
\value{
A ggplot based graphic
}
\description{
Title
}
