% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plsda.R
\name{plot_plsda}
\alias{plot_plsda}
\title{Plot PLS-DA using mixOmics \code{plsda}}
\usage{
plot_plsda(
  res.plsda,
  comp = c(1, 2),
  color_vector = tol21rainbow,
  shape,
  top.loads = 10
)
}
\arguments{
\item{res.plsda}{An object returned by \code{\link[mixOmics]{plsda}}.}

\item{comp}{Which component to plot.}

\item{color_vector}{Color vector}

\item{shape}{Shape of the dots.}

\item{top.loads}{Top N features to plot. Default = 10.
Loads are defined by the absolute value of each features for each components. If \code{top.loads = 10} the function
will keep the top10 features of component 1 and the top 10 features of the component 2.}
}
\description{
A ggplot2 based function to plot PLS-DA based on the object returned by \code{\link[mixOmics]{plsda}}.
The function will also plot the features loadings, a biplot then. This function can also handle results of a \code{splsda}.
}
\examples{

data(metabolomic)
res= mixOmics::plsda(X = metabolomic[,5:217], Y = metabolomic$birth_type, ncomp=5)

# Basic plotting
plot_plsda(res, comp= c(1,2), shape=21, top.loads= 20)

# Improve a bit the plot.
plot_plsda(res, comp= c(1,2), shape=21, top.loads= 20, color_vector= tol21rainbow[c(2,14)])+
guides(fill= guide_legend("Birth route"))+
theme_bw()+
theme(legend.position = "bottom",
       axis.title = element_text(size=15, face="bold"),
       panel.grid = element_blank(),
       legend.text = element_text(size=15),
       legend.title = element_text(size=20, face="bold"),
       axis.text = element_text(size=15))


}
