% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qual.R
\name{plot_qual}
\alias{plot_qual}
\title{Check the length of your reads.}
\usage{
plot_qual(list, n = 5, color_vector = NULL, project_names)
}
\arguments{
\item{list}{An object returned by \link{list_fastq}}

\item{n}{Number of fastq you used in summarise_fastq()}

\item{color_vector}{A color vector}
}
\description{
Check the length of your reads.
}
