% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reduction.R
\name{plot_reduction}
\alias{plot_reduction}
\title{Plot reduction}
\usage{
plot_reduction(
  mat,
  axis_x = 1,
  axis_y = 2,
  nf = 5,
  method = "PCA",
  group = NULL,
  permanova = F,
  color_vector = c("cyan4", "brown", "deepskyblue", "black", "red"),
  legend_title = NULL,
  title = NULL,
  lwd = 1,
  conf = 0.9,
  cex = 2,
  font = 2,
  pch = 20,
  draw = "lines",
  ylimits = "auto",
  xlimits = "auto",
  text = F,
  ncol = 1,
  species = F,
  x.intersp = 1,
  y.intersp = 0.5,
  where = "topleft",
  inset = 0.2,
  boxplot = T,
  pca = T,
  ...
)
}
\arguments{
\item{mat}{A matrix with samples in rows}

\item{axis_x}{Number of the component for X axis}

\item{axis_y}{Number of the component for Y axis}

\item{nf}{Number of component to calculate}

\item{method}{Method of reduction to use. Currently available : "PCA", "CA", "DCA", "tsne"}

\item{group}{The group to colorise and put ellipse}

\item{permanova}{Logical}

\item{color_vector}{Color vector}

\item{legend_title}{Legend title}

\item{title}{Title}

\item{lwd}{Line width}

\item{conf}{Confidence interval. Default = 0.9.}

\item{cex}{Size of the text}

\item{font}{Type of font. Default = 2 (bold).}

\item{pch}{Shape}

\item{draw}{Categorical, draw ellipses as lines or as polygons}

\item{ylimits}{@param xlimits Limits for the axis}

\item{ncol}{Number of columns for the legend. Default = 1}

\item{x.intersp}{Adjust the legend: character interspacing factor for horizontal (x)
spacing between symbol and legend text.
@param y.intersp Adjust the legend: vertical (y) distances (in lines of text shared above/below each legend entry). A vector with one element for each row of the legend can be used.}

\item{where}{Position for the permanova value. Default = "topleft"}

\item{inset}{Adjust the legend: inset distance(s) from the margins as a fraction of the plot region when legend is placed by keyword.}

\item{boxplot}{Logical. Plot the marginal boxplots or not. Default= T}

\item{pca}{Logical. For tsne only. Does the tsne need to be run on a PCA first ?}

\item{...}{}
}
\description{
Plot reduction
}
