% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Plot volcano}
\usage{
plot_volcano(volcano, size = 2, col = c("#771155", "#77CCCC", "grey"), ...)
}
\arguments{
\item{volcano}{An object returned by the \link{fold_change} function}

\item{col}{Color vector.}

\item{...}{Addtionnal arguments passed to \link{geom_point(aes(...))}}
}
\description{
Make a volcano plot
}
