% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xy_variance.R
\name{plot_xy_variance}
\alias{plot_xy_variance}
\title{Plot the variance for two factors.}
\usage{
plot_xy_variance(
  variance,
  x,
  y,
  fac,
  col = c("brown", "orange", "grey"),
  max.overlap = 20
)
}
\arguments{
\item{variance}{An object returned by \link{calculate_variance}.}

\item{x}{Factor to be plotted as x.}

\item{y}{Factor to be plotted as y.}

\item{fac}{The factor to depict p.values. Only the points with p.adj < 0.05
and p < 0.05 will be colorized and labelled}

\item{col}{Color vector.}
}
\description{
Plot the variance for two factors.
}
