% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_models.R
\name{predict_models}
\alias{predict_models}
\title{Predict the classification based on the model training.}
\usage{
predict_models(model_obj)
}
\arguments{
\item{model_obj}{A object returned by \code{\link[=screen_model]{screen_model()}}.}
}
\value{
A list of predicted values and the confusion matrix for each variables.
}
\description{
Predict the classification based on the model training.
}
\examples{
TBD
}
