% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_check.R
\name{qc_check}
\alias{qc_check}
\title{A wrapper function to plot the quality of fastq.}
\usage{
qc_check(fastq_list, n = 30, save_plot = F)
}
\arguments{
\item{fastq_list}{An object given by list_fastq()}

\item{n}{Number of fastq you want to incorpore}

\item{save_plot}{Save the plot as .png plot or not. Default= FALSE}
}
\value{
A ggplot2 object.
}
\description{
A wrapper function to plot the quality of fastq.
}
\examples{
No example.
}
