% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng_fastq.R
\name{rng_fastq}
\alias{rng_fastq}
\title{Randomly select fastq files}
\usage{
rng_fastq(fastq_list, n = 10)
}
\arguments{
\item{fastq_list}{List of fastq returned by \code{\link[vegan]{list_fastq}}}

\item{n}{Number of fastq to select}
}
\description{
Randomly select fastq files
}
