% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_models.R
\name{screen_models}
\alias{screen_models}
\title{Title Screen a model on all your variables.}
\usage{
screen_models(
  obj,
  model = "plsda",
  cores = 1,
  ncomp = 10,
  number = 10,
  repeats = 3,
  prop = 0.5,
  train = T,
  ...
)
}
\arguments{
\item{obj}{An object of class phyloseq or a list of two levels containing a matrix of values and a dataframe of variables.}

\item{model}{The model tested, for now this function supports \code{\link[randomForest:randomForest]{randomForest::randomForest()}}, \code{\link[glmnet:glmnet]{glmnet::glmnet()}} and \code{\link[caret:plsda]{caret::plsda()}}.}

\item{cores}{The number of cores you want to use.}

\item{ncomp}{Number of component you want to keep for \code{\link[caret:plsda]{caret::plsda()}}. Default =10.}

\item{number}{Either the number of folds or number of resampling iterations}

\item{repeats}{For repeated k-fold cross-validation only: the number of complete sets of folds to compute. Default = 3}

\item{prop}{The proportion of the variables to use in the training model. Default = 0.5.}

\item{train}{Wheter or not you want to train the models in order to predict.
If you select FALSE the function will use all the data to perform classification or find
the best features that explain your data according to the variables. Default=T.}

\item{...}{Argument passed to \code{\link[caret:train]{caret::train()}}.}
}
\value{
A list containing
\item{model}{A \code{\link[caret:train]{caret::train()}} class}
\item{Y}{The trimmed variables}
\item{X}{The trimmed values}
\item{train_y}{The variables used to train the model}
\item{train_x}{The values used to train the model}
\item{pred_x}{The predicted values}
\item{pred_y}{The predicted classification}
}
\description{
A function to test machine learning on your different variables of the metadata.
This function supports phyloseq object as input or a list containing a values matrix and a variable dataframe.
This function is based on \link{caret}.
}
\examples{
screen_models(enterotype, model="glmnet", prop=0.5, ncomp=20)

screen_models(enterotype, model="plsda", cores=1, ncomp=10, number=10, repats=5, prop=0.6, train=T)
}
