% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_fastq.R
\name{select_fastq}
\alias{select_fastq}
\title{Select the fastq in the directory you want.}
\usage{
select_fastq(names_list, fastq_list)
}
\arguments{
\item{names_list}{A list of names}

\item{fastq_list}{A list of fastq}
}
\value{
A list.
}
\description{
Select the fastq in the directory you want.
}
\examples{
No example
}
