% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_table.R
\name{seq_table}
\alias{seq_table}
\title{IgASeq table}
\usage{
seq_table(physeq, sample_name, sorting_names, cols_to_keep = "all")
}
\arguments{
\item{physeq}{A phyloseq object to analyse your data. (\code{\link{phyloseq-class}})}

\item{sample_name}{Character vector to identify the unique sample name from which you performed
the sortings}

\item{sorting_names}{Character vector containing the variable for the sorting fractions.}

\item{cols_to_keep}{The columns of \code{\link{phyloseq::sample_data-class}}. Default = "all", better to keep all of the columns.}
}
\value{
A list equal to the length of unique samples that originated the sorted samples. The list contain the taxa collapsed to a single column (# as separator),
the sample_data() from the given phyloseq object and the pos, neg1 and neg2 relative abundance.
}
\description{
Prepare phyloseq object to perform the slide_z function.
}
