% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_z_modern.R
\name{slide_z_modern}
\alias{slide_z_modern}
\title{Modern version of the Z test on the IgASeq data.}
\usage{
slide_z_modern(df, log2_ratio, log10_abundance, deltaX = deltaX, alpha = alpha)
}
\arguments{
\item{df}{The dataframe returned by \code{\link{seq_table}}}

\item{log2_ratio}{The \code{\link{log_ratio}} of neg}

\item{log10_abundance}{The}

\item{deltaX}{The}

\item{alpha}{The}
}
\value{
A slidez
}
\description{
\code{slide_z_modern} will use the dispersion between the two negative fractions to mesure the technical dispersion and apply the Z stat on the pos/neg1 ration
}
\examples{
No example
}
