% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_plot.R
\name{ternary_plot}
\alias{ternary_plot}
\title{Plot fold change or abundance as a ternary plot}
\usage{
ternary_plot(
  fold,
  adjusted.p.val = T,
  measure = "fold.change",
  stat = "median"
)
}
\arguments{
\item{fold}{A dataframe form the object returned by \link{fold_change} function.
Be aware that you need to specify which dataframe you want or use a \code{lapply} like approach. See examples for more.}

\item{adjusted.p.val}{Should the function filter features base on adjusted p values or not. Default =T}

\item{measure}{Should the function use fold change measure or the mean measures. Default = "fold.change".}

\item{stat}{Not implemented yet.}
}
\description{
A wrapper function using \link{ggtern} function to produce a nice ternary plot.
}
