% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_dispersion.R
\name{beta_dispersion}
\alias{beta_dispersion}
\title{This a wrapper function to plot beta diversity directly from a phyloseq
object. This is based on the base package and mostly \link{vegan}.}
\usage{
beta_dispersion(
  physeq,
  axis_x = 1,
  axis_y = 2,
  dist = "bray",
  nf = 5,
  method = "PCoA",
  group = NULL,
  color_vector = c("cyan4", "brown", "deepskyblue", "black", "red"),
  legend_title = NULL,
  lwd = 1,
  conf = 0.9,
  cex = 2,
  font = 2,
  pch = 20,
  draw = "lines",
  ylimits = "auto",
  xlimits = "auto",
  text = F,
  ncol = 1,
  species = F,
  x.intersp = 1,
  y.intersp = 0.5,
  permanova = F,
  where = "topleft",
  inset = 0.2,
  ...
)
}
\arguments{
\item{physeq}{A phyloseq object.}

\item{dist}{Either a distance mecthod given as string or a distance object.
Default= "bray".}

\item{nf}{Number of component to keep. Default= 5.}

\item{method}{Method use to plot the beta diversity. Currently this function
supports only "PCoA".}

\item{group}{String defining the groups you want to plot on the beta diversity.}

\item{color_vector}{Optionnal. A color vector you wish to use for plotting}

\item{legend_title}{Optionnal. A legend title.}

\item{lwd, cex, font, pch, ylimits, xlimits}{Parameters imported from base plotting}

\item{conf}{Interval of confidence for the ellipses. Default= 0.9.}

\item{draw}{Either draw lines or polygon. Default = "lines".}

\item{ncol}{Number of columns for the legend. Default = 1.}

\item{x.intersp}{character interspacing factor for horizontal (x) spacing between symbol and legend text.}

\item{y.intersp}{vertical (y) distances (in lines of text shared above/below each legend entry). A vector with one element for each row of the legend can be used.}

\item{permanova}{Make the permanova analysis. Default = FALSE.}

\item{where}{Where you want the permanova result.}

\item{inset}{Permanova legend adjustment see \link{legend}}

\item{...}{Passed to other packages.}

\item{textOptionnal.}{Labels to use for plotting. Default will be the name of the groups you gave.}
}
\description{
This a wrapper function to plot beta diversity directly from a phyloseq
object. This is based on the base package and mostly \link{vegan}.
}
