% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_dispersion.R
\name{beta_dispersion}
\alias{beta_dispersion}
\title{Make an unconstrained analysis of beta diversity.}
\usage{
beta_dispersion(
  physeq,
  axis_x = 1,
  axis_y = 2,
  dist = "bray",
  nf = 5,
  method = "PCoA",
  type = "boxplot",
  group = NULL,
  stat = "none",
  color_vector = c("cyan4", "brown", "deepskyblue", "black", "red"),
  legend_title = NULL,
  lwd = 1,
  conf = 0.9,
  cex = 2,
  font = 2,
  pch = 20,
  draw = "lines",
  ylimits = "auto",
  xlimits = "auto",
  text = F,
  ncol = 1,
  species = F,
  x.intersp = 1,
  y.intersp = 0.5,
  where = "topleft",
  inset = 0.2,
  pca = T,
  stat.cex = 2,
  ...
)
}
\arguments{
\item{physeq}{A phyloseq object.}

\item{axis_x}{Which component for axis x. Default = 1}

\item{axis_y}{Which component for axis y. Default = 2}

\item{dist}{Either a distance mecthod given as string or a distance object.
Default= "bray".}

\item{nf}{Number of component to keep. Default= 5.}

\item{method}{Method use to plot the beta diversity. Currently this function
supports only "PCoA".}

\item{type}{Chose the type of plot you want, choices are : "boxplot", "pure" or "arrows". Default="boxplot".}

\item{group}{String defining the groups you want to plot on the beta diversity}

\item{stat}{Either permanova \code{\link[vegan]{adonis2}} with default parameter or \code{\link[vegan]{envfit}} function.}

\item{color_vector}{Optionnal. A color vector you wish to use for plotting}

\item{legend_title}{Optionnal. A legend title.}

\item{lwd}{Line width}

\item{conf}{Interval of confidence for the ellipses. Default= 0.9.}

\item{cex}{Text size}

\item{font}{Font type}

\item{pch}{Shape type}

\item{draw}{Either draw lines or polygon. Default = "lines".}

\item{ylimits}{Limits for Y axis, if not specified they will be automatically set}

\item{xlimits}{Limits for X axis, if not specified they will be automatically set}

\item{text}{Should the labels be printed}

\item{ncol}{Number of columns for the legend. Default = 1.}

\item{species}{NMDS only, will plot the species loadings on the graph.}

\item{x.intersp}{character interspacing factor for horizontal (x) spacing between symbol and legend text.}

\item{y.intersp}{vertical (y) distances (in lines of text shared above/below each legend entry). A vector with one element for each row of the legend can be used.}

\item{where}{Where you want the permanova result.}

\item{inset}{Permanova legend adjustment see \link{legend}}

\item{pca}{Argument for t-SNE, does the t-SNE needs to run on a PCA first or not ? Default=FAL}

\item{stat.cex}{ize of the stats text.}

\item{...}{}
}
\value{
A plot rendered using the \code{base} package and vegan functions :
\itemize{
\item \code{\link[vegan]{ordispider}}
\item \code{\link[vegan]{ordiellipse}}
}
}
\description{
This a wrapper function to plot beta diversity directly from a phyloseq
object. This is based on the base package and mostly \link{vegan}.
}
\examples{

data(enterotype)
# PCoA based on
beta_dispersion(enterotype, dist = "bray", method = "PCoA", group = "SeqTech",
color_vector = c("#777711", "#117777", "#DD7788"),
legend_title = "Sequencing tech", lwd = 2,
font = 2, draw = "polygon", text = T, permanova = T,
y.intersp = 0.7)

}
\seealso{
\code{\link[=plot_reduction]{plot_reduction()}}
\code{\link[=plot_constrained_reduction]{plot_constrained_reduction()}}
\code{\link[=constrained_beta_dispersion]{constrained_beta_dispersion()}}
}
