% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_variance.R
\name{calculate_variance}
\alias{calculate_variance}
\title{Calculate the variance for each factors.}
\usage{
calculate_variance(meta, feat, cores = 1)
}
\arguments{
\item{meta}{Metadata containing the factors to be used.}

\item{feat}{Features matrix}
}
\description{
Calculate the variance for each factors.
}
