% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_length.R
\name{check_length}
\alias{check_length}
\title{Check the length of reads in fastq.}
\usage{
check_length(fastq, names, cores = 1)
}
\arguments{
\item{fastq}{To do}

\item{names}{TBD}

\item{cores}{TBD}
}
\value{
TBD
}
\description{
Check the length of reads in fastq.
}
\examples{
No example
}
