% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_primer.R
\name{find_primer}
\alias{find_primer}
\title{Title}
\usage{
find_primer(
  fastq_list,
  fwd_primer,
  rv_primer,
  max.mismatch,
  allow.indels = FALSE,
  pair_end = T,
  fixed = F
)
}
\arguments{
\item{fixed}{}
}
\description{
Title
}
