% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change.R
\name{fold_change}
\alias{fold_change}
\title{Calculate fold change and p.value for your data.}
\usage{
fold_change(meta, feat, cores = 2)
}
\arguments{
\item{meta}{A metadata containing the factors you want to test.}

\item{feat}{A features/ASV table with the sample in rows.}

\item{cores}{Number of cores you want to use. Default = 1.}
}
\description{
Accepts a metadata and a feature/microbiome matrix.
}
