% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_and_remove_chimera.R
\name{merge_and_remove_chim}
\alias{merge_and_remove_chim}
\title{Merge reads and remove chimera}
\usage{
merge_and_remove_chim(dd, filt_fastq_list, cores = 1, ...)
}
\arguments{
\item{dd}{Object returned by \link{dereplicate} or \link{dada}.}

\item{filt_fastq_list}{The list of filtered fastq}

\item{cores}{Number of cores. Default=1}

\item{...}{Parameter passed to \link{removeBimeraDenovo}}
}
\value{
Return a count matrix.
}
\description{
Merge reads and remove chimera
}
\examples{
No example.
}
