% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_constrained_reduction.R
\name{plot_constrained_reduction}
\alias{plot_constrained_reduction}
\title{Make a constrained reduction on matrices and data.frame.}
\usage{
plot_constrained_reduction(
  mat,
  clinical_data,
  axis_x = 1,
  axis_y = 2,
  model,
  nf = 5,
  method = "CCA",
  type = "boxplot",
  group = NULL,
  stat = "none",
  color_vector = c("cyan4", "brown", "deepskyblue", "black", "red"),
  legend_title = NULL,
  lwd = 1,
  conf = 0.9,
  cex = 2,
  font = 2,
  pch = 20,
  draw = "lines",
  ylimits = "auto",
  xlimits = "auto",
  text = F,
  ncol = 1,
  x.intersp = 1,
  y.intersp = 0.5,
  where = "topleft",
  inset = 0.2,
  stat.cex = 2,
  ...
)
}
\arguments{
\item{mat}{A numeric matrix with samples as rows}

\item{clinical_data}{An integer specifying the columns that are discrete values.}

\item{axis_x}{Which component for axis x. Default = 1}

\item{axis_y}{Which component for axis y. Default = 2}

\item{model}{Specify here the model to be used for regression}

\item{nf}{Number of components}

\item{method}{The mathematical method to reduce the dimensions. Currently this function supports CCA, RDA and dbRDA}

\item{type}{Chose the type of plot you want, choices are : "boxplot", "pure" or "arrows". Default="boxplot".}

\item{group}{String defining the groups you want to plot on the beta diversity.}

\item{stat}{Either permanova \code{\link[vegan]{adonis2}} with default parameter or \code{\link[vegan]{envfit}} function.}

\item{color_vector}{Optionnal. A color vector you wish to use for plotting}

\item{legend_title}{Optionnal. A legend title.}

\item{lwd}{Line width}

\item{conf}{Confidence interval. Default = 0.9.}

\item{cex}{Text size}

\item{font}{Font type}

\item{pch}{Shape type}

\item{draw}{Categorical, draw ellipses as lines or as polygons#'}

\item{ylimits}{Limits for Y axis, if not specified they will be automatically set}

\item{xlimits}{Limits for X axis, if not specified they will be automatically set}

\item{text}{Plot labels or not.}

\item{ncol}{Number of columns for the legend. Default = 1.}

\item{x.intersp}{character interspacing factor for horizontal (x) spacing between symbol and legend text.}

\item{y.intersp}{vertical (y) distances (in lines of text shared above/below each legend entry). A vector with one element for each row of the legend can be used.}

\item{where}{Where you want the permanova result.}

\item{inset}{Permanova legend adjustment see \link{legend}}

\item{stat.cex}{Size of the stats text.}

\item{...}{}
}
\value{
A plot rendered using the \code{base} package and vegan functions :
\itemize{
\item \code{\link[vegan]{ordispider}}
\item \code{\link[vegan]{ordiellipse}}
}
}
\description{
Currently the function supports only the constrained reduction around categorical data.
It's not possible yet to give two different Omics datasets to find correlation between them.
}
\examples{



data(metabolomic)

# RDA based on metabolomic data and constrained to breastfeeding status and gender.

res= plot_constrained_reduction( metabolomic, clinical_data = 1:5,
                                 stat="envfit", method = "RDA", model = "breastfeeding+sex",
                                 group = "birth_type", type="arrows")

# CCA
res = plot_constrained_reduction(metabolomic,  clinical_data = 1:4, stat="permanova", method = "RDA", model = "breastfeeding+sex",
                                 group = "birth_type", type="arrows")


# dbRDA
res = plot_constrained_reduction(metabolomic, clinical_data = 1:4, stat="envfit", method = "CCA", model = "breastfeeding+sex",
                                 group = "birth_type", type="arrows")
anova(res, by="term")
}
\seealso{
\code{\link[=plot_reduction]{plot_reduction()}}
\code{\link[=plot_constrained_reduction]{plot_constrained_reduction()}}
\code{\link[=beta_dispersion]{beta_dispersion()}}
}
