% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Plot volcano}
\usage{
plot_volcano(volcano, size = 2, col = c("#771155", "#77CCCC", "grey"), ...)
}
\arguments{
\item{volcano}{An object returned by the \link{fold_change} function}

\item{col}{Color vector.}

\item{...}{Addtionnal arguments passed to \link{geom_point(aes(...))}}
}
\value{
A ggplot2 object with features labelled depending on the p-values and the corrected p-values.
}
\description{
Make a volcano plot based on the object returned by \link{fold_change}
}
\examples{
fold= fold_change(clinical_data = metabolomic[,1:4], metabolomic[,6:217], cores = 10)
lapply(fold, plot_volcano)
plot_volcano(fold$MeCT, size=2)
}
