% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xy_variance.R
\name{plot_xy_variance}
\alias{plot_xy_variance}
\title{Plot the variance for two factors.}
\usage{
plot_xy_variance(
  variance,
  x,
  y,
  corrected = F,
  col_vector = c("purple", "orange", "brown", "grey"),
  max.overlap = 20,
  size = 3
)
}
\arguments{
\item{variance}{An object returned by \link{calculate_variance}.}

\item{x}{Factor to be plotted as x.}

\item{y}{Factor to be plotted as y.}

\item{corrected}{Use p value adjusted by FDR or not. Default=F.}

\item{col}{Color vector.}
}
\description{
Plot the variance for two factors.
}
